v0.7.0-refactor: Move Transformer modules to nn.
