+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "symm-learning"
-version = "0.6.4"
+version = "0.7.0"
 
 description = "Library of torch modules and utilities of equivariant/invariant learning"
 readme = "README.md"
