+++ test/test_modules.py
@@ -77,6 +77,173 @@ def test_change2disentangled(group: Group):  # noqa: D103
         pytest.param(Icosahedral(), id="icosahedral"),
     ],
 )
+@pytest.mark.parametrize("mx", [2])
+@pytest.mark.parametrize("num_heads", [1, 2])
+@pytest.mark.parametrize("num_layers", [1, 5])
+def test_etransformer_decoder(group: Group, mx: int, num_heads: int, num_layers: int):
+    """Check equivariance and fast inference consistency of eTransformerDecoderLayer."""
+    from symm_learning.nn import eTransformerDecoderLayer
+
+    G = group
+    rep = direct_sum([G.regular_representation] * mx)
+
+    decoder_kwargs = dict(
+        in_rep=rep,
+        nhead=num_heads,
+        dim_feedforward=rep.size * 4,
+        dropout=0.0,  # dropout=0 for train/eval consistency
+        activation="relu",
+        norm_first=True,
+        batch_first=True,
+        norm_module="rmsnorm",
+        bias=True,
+    )
+
+    # Create single layer or stacked layers
+    if num_layers == 1:
+        decoder = eTransformerDecoderLayer(**decoder_kwargs)
+    else:
+        base_layer = eTransformerDecoderLayer(**decoder_kwargs)
+        decoder = torch.nn.TransformerDecoder(decoder_layer=base_layer, num_layers=num_layers)
+
+    # Equivariance check
+    decoder.eval()
+    if num_layers == 1:
+        decoder.check_equivariance(atol=1e-4, rtol=1e-4)
+    else:
+        # For stacked layers, use manual equivariance check
+        def act(rep_local, g, tensor):
+            mat = torch.tensor(rep_local(g), dtype=tensor.dtype, device=tensor.device)
+            return torch.einsum("ij,...j->...i", mat, tensor)
+
+        B, tgt_len, mem_len = 3, 2, 3
+        for _ in range(5):
+            g = G.sample()
+            tgt = torch.randn(B, tgt_len, rep.size)
+            mem = torch.randn(B, mem_len, rep.size)
+            out = decoder(tgt=tgt, memory=mem)
+            g_out = decoder(tgt=act(rep, g, tgt), memory=act(rep, g, mem))
+            g_out_exp = act(rep, g, out)
+            assert torch.allclose(g_out, g_out_exp, atol=1e-3, rtol=1e-3), (
+                f"Decoder stack equivariance failed, max err {(g_out - g_out_exp).abs().max().item():.3e}"
+            )
+
+    # Fast inference consistency test
+    B, tgt_len, mem_len = 4, 3, 5
+    tgt = torch.randn(B, tgt_len, rep.size)
+    mem = torch.randn(B, mem_len, rep.size)
+
+    # 1. Update weights with some arbitrary loss
+    decoder.train()
+    decoder.zero_grad()
+    y = decoder(tgt, mem) if num_layers == 1 else decoder(tgt=tgt, memory=mem)
+    target = torch.randn_like(y)
+    loss = torch.nn.functional.mse_loss(y, target)
+    loss.backward()
+    with torch.no_grad():
+        for p in decoder.parameters():
+            if p.grad is not None:
+                p -= 0.1 * p.grad
+
+    # 2. Forward in train mode with updated weights
+    decoder.zero_grad()
+    y_train = decoder(tgt, mem) if num_layers == 1 else decoder(tgt=tgt, memory=mem)
+
+    # 3. Forward in eval mode
+    decoder.eval()
+    y_eval = decoder(tgt, mem) if num_layers == 1 else decoder(tgt=tgt, memory=mem)
+
+    assert torch.allclose(y_train, y_eval, atol=1e-5, rtol=1e-5), (
+        f"eTransformerDecoder output in eval mode must match train mode (layers={num_layers})"
+    )
+
+
+@pytest.mark.parametrize(
+    "group",
+    [
+        pytest.param(CyclicGroup(5), id="cyclic5"),
+        pytest.param(Icosahedral(), id="icosahedral"),
+    ],
+)
+@pytest.mark.parametrize("mx", [2])
+@pytest.mark.parametrize("num_heads", [1, 2])
+@pytest.mark.parametrize("num_layers", [1, 5])
+def test_etransformer_encoder(group: Group, mx: int, num_heads: int, num_layers: int):
+    """Check equivariance and fast inference consistency of eTransformerEncoderLayer."""
+    from symm_learning.nn.transformer.etransformer import eTransformerEncoderLayer
+
+    G = group
+    rep = direct_sum([G.regular_representation] * mx)
+
+    encoder_kwargs = dict(
+        in_rep=rep,
+        nhead=num_heads,
+        dim_feedforward=rep.size * 4,
+        dropout=0.0,  # dropout=0 for train/eval consistency
+        activation="relu",
+        norm_first=True,
+        batch_first=True,
+        norm_module="rmsnorm",
+        bias=True,
+    )
+
+    # Create single layer or stacked layers
+    if num_layers == 1:
+        encoder = eTransformerEncoderLayer(**encoder_kwargs)
+    else:
+        base_layer = eTransformerEncoderLayer(**encoder_kwargs)
+        encoder = torch.nn.TransformerEncoder(
+            encoder_layer=base_layer, num_layers=num_layers, enable_nested_tensor=False
+        )
+
+    # Equivariance check
+    encoder.eval()
+    check_equivariance(
+        encoder,
+        input_dim=3,
+        module_name=f"eTransformerEncoder(layers={num_layers})",
+        in_rep=rep,
+        out_rep=rep,
+        atol=1e-4,
+        rtol=1e-4,
+    )
+
+    # Fast inference consistency test
+    B, L = 4, 5
+    x = torch.randn(B, L, rep.size)
+
+    # 1. Update weights with some arbitrary loss
+    encoder.train()
+    encoder.zero_grad()
+    y = encoder(x)
+    target = torch.randn_like(y)
+    loss = torch.nn.functional.mse_loss(y, target)
+    loss.backward()
+    with torch.no_grad():
+        for p in encoder.parameters():
+            if p.grad is not None:
+                p -= 0.1 * p.grad
+
+    # 2. Forward in train mode with updated weights
+    encoder.zero_grad()
+    y_train = encoder(x)
+
+    # 3. Forward in eval mode
+    encoder.eval()
+    y_eval = encoder(x)
+
+    assert torch.allclose(y_train, y_eval, atol=1e-5, rtol=1e-5), (
+        f"eTransformerEncoder output in eval mode must match train mode (layers={num_layers})"
+    )
+
+
+@pytest.mark.parametrize(
+    "group",
+    [
+        pytest.param(CyclicGroup(5), id="cyclic5"),
+        pytest.param(Icosahedral(), id="icosahedral"),
+    ],
+)
 @pytest.mark.parametrize("mx", [3])
 @pytest.mark.parametrize("my", [3, 5])
 def test_equiv_multivariate_normal(group: Group, mx: int, my: int):
